requirements.txt
Django==5.2
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-ckeditor==6.7.0
djangorestframework==3.14.0
python-decouple==3.8
whitenoise==6.6.0
requests==2.31.0
Pillow==10.2.0
django>=5.0.0
requests>=2.31.0
python-decouple>=3.8
urllib3>=2.1.0
certifi>=2024.2.2
pyOpenSSL>=23.2.0
cryptography>=41.0.0
idna>=3.4
dnspython>=2.6.0